The RF COMS are Functional

+++ Arduino Sketches/CalculateCentre/CALCENMK-I/CALCENMK-I.ino
@@ -1,4 +1,4 @@
-double cenlat, cenlon;
+float cenlat, cenlon;
 void setup() {
   // put your setup code here, to run once:
